+++ library/src/main/java/io/sfrei/tracksearch/tracks/YouTubeTrack.java
@@ -1,19 +1,24 @@
 package io.sfrei.tracksearch.tracks;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import io.sfrei.tracksearch.tracks.deserializer.YouTubeTrackDeserializer;
 import io.sfrei.tracksearch.clients.setup.TrackSource;
+import io.sfrei.tracksearch.tracks.deserializer.YouTubeTrackDeserializer;
 import io.sfrei.tracksearch.tracks.metadata.YouTubeTrackInfo;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.util.function.Function;
 
-@Getter
-@Setter
+
 @JsonDeserialize(using = YouTubeTrackDeserializer.class)
 public class YouTubeTrack extends BaseTrack implements Track {
 
+    @Getter
+    @Setter
     private YouTubeTrackInfo trackInfo;
+
+    @Setter
+    private Function<YouTubeTrack, String> streamUrlProvider;
 
     public YouTubeTrack(String title, Long length, String url) {
         super(TrackSource.Youtube, title, length, url);
@@ -24,4 +29,8 @@ public class YouTubeTrack extends BaseTrack implements Track {
         return this.trackInfo;
     }
 
+    @Override
+    public String getStreamUrl() {
+        return streamUrlProvider.apply(this);
+    }
 }